# public/swagger/user.json
{
    "swagger": "2.0",
    "info": {
        "description": "All APIs require JWT header",
        "version": "1.0.0",
        "title": "Venn user service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "support@getvenn.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:2331",
    "basePath": "/v1",
...
    "schemes": [
        "http"
    ],
    "paths":{
...
        "/user/profile": {
            "post" : {
                "summary": "Profile",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "profile",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/Profile"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
        "/user/publicprofile": {
            "post" : {
                "summary": "Public Profile",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "PublicProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Public Profile",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/PublicProfile"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            },
            "get" :{
                "summary": "Get Public Profile",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "GetPublicProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            },
            "put": {
                "summary": "Update Public Profile",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "UpdatePublicProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Update Public Profile",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/PublicProfile"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
...
        "/user/report": {
            "post":{
                "summary": "User Report",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "ReportUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User Report",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/Report"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
...
        "/user/updateUserQuickBio": {
            "put": {
                "summary": "Update User Quick Bio",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "UpdateUserQuickBio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Update User Quick Bio",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/UpdateUserQuickBio"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
        "/user/updateUserSocialLink": {
            "put": {
                "summary": "Update User Social Link",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "UpdateUserSocialLink",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Update User SocialLink",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/UpdateUserSocialLink"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
        "/user/changePassword":{
            "put":{
                "summary": "Change Password",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "ChangePassword",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Change Password",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/ChangePassword"
                        }
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ] 
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}